requirements.txt
asgiref==3.7.2
cffi==1.16.0
contextlib2==21.6.0
crispy-bootstrap5==0.7
cryptography==41.0.5
Django==4.2.6
django-crispy-forms==2.1
django-ninja==0.22.2
django-ninja-extra==0.19.8
django-ninja-jwt==5.2.6
injector==0.21.0
psycopg2==2.9.9
pycparser==2.21
pydantic==1.10.13
PyJWT==2.8.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3